User error checks

+++ backend/src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {ConflictException, Injectable, NotFoundException} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserRepository } from 'src/user/user.repository';
@@ -40,17 +40,27 @@ export class UserService {
   }
 
   async createUser(createUserDto: CreateUserDto): Promise<User> {
+    const userCheck = this.userRepository.findByEmail(createUserDto.email);
+    if (userCheck) {
+      throw new ConflictException('User with such email already exists');
+    }
     return this.userRepository.create(createUserDto);
   }
 
   async updateUser(id: number, updateUserDto: UpdateUserDto): Promise<User> {
     const existingUser = await this.findById(id);
+    if (!existingUser) {
+      throw new NotFoundException('User not found');
+    }
     const updatedUser = { ...existingUser, ...updateUserDto };
     return this.userRepository.update(id, updatedUser);
   }
 
   async deleteUser(id: number): Promise<void> {
     const existingUser = await this.findById(id);
+    if (!existingUser) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
     await this.userRepository.delete(existingUser.id);
   }
 
@@ -69,6 +79,9 @@ export class UserService {
 
   async deleteUserTask(userId: number, taskId: number): Promise<Task> {
     const user = await this.userRepository.findById(userId);
+    if (!user) {
+      throw new NotFoundException(`User with ID ${userId} not found`);
+    }
     const task = await this.taskService.findById(taskId);
     return await this.userRepository.deleteTaskFromUser(user, task);
   }
@@ -108,6 +121,9 @@ export class UserService {
 
   async getSortedUserTasks(userId: number): Promise<Task[]> {
     const user = await this.findById(userId);
+    if (!user) {
+      throw new Error(`User with ID ${userId} not found`);
+    }
     // Sort tasks by calculated priority in descending order
     return user.tasks.sort(
       (a, b) =>
